container: Separate Unix tests
