+++ container_test.go
@@ -15,14 +15,11 @@ import (
 	"net/http/httptest"
 	"net/url"
 	"os"
-	"path/filepath"
 	"reflect"
 	"strconv"
 	"strings"
 	"testing"
 	"time"
-
-	"github.com/hashicorp/go-cleanhttp"
 )
 
 func TestStateString(t *testing.T) {
@@ -1717,37 +1714,6 @@ func TestExportContainer(t *testing.T) {
 	}
 }
 
-func TestExportContainerViaUnixSocket(t *testing.T) {
-	content := "exported container tar content"
-	var buf []byte
-	out := bytes.NewBuffer(buf)
-	tempSocket := tempfile("export_socket")
-	defer os.Remove(tempSocket)
-	endpoint := "unix://" + tempSocket
-	u, _ := parseEndpoint(endpoint, false)
-	client := Client{
-		HTTPClient:             cleanhttp.DefaultClient(),
-		Dialer:                 &net.Dialer{},
-		endpoint:               endpoint,
-		endpointURL:            u,
-		SkipServerVersionCheck: true,
-	}
-	listening := make(chan string)
-	done := make(chan int)
-	containerID := "4fa6e0f0c678"
-	go runStreamConnServer(t, "unix", tempSocket, listening, done, containerID)
-	<-listening // wait for server to start
-	opts := ExportContainerOptions{ID: containerID, OutputStream: out}
-	err := client.ExportContainer(opts)
-	<-done // make sure server stopped
-	if err != nil {
-		t.Errorf("ExportContainer: caugh error %#v while exporting container, expected nil", err.Error())
-	}
-	if out.String() != content {
-		t.Errorf("ExportContainer: wrong stdout. Want %#v. Got %#v.", content, out.String())
-	}
-}
-
 func runStreamConnServer(t *testing.T, network, laddr string, listening chan<- string, done chan<- int, containerID string) {
 	defer close(done)
 	l, err := net.Listen(network, laddr)
@@ -2016,63 +1982,6 @@ func TestTopContainerWithPsArgs(t *testing.T) {
 	expectedURI := "/containers/abef348/top?ps_args=aux"
 	if !strings.HasSuffix(fakeRT.requests[0].URL.String(), expectedURI) {
 		t.Errorf("TopContainer: Expected URI to have %q. Got %q.", expectedURI, fakeRT.requests[0].URL.String())
-	}
-}
-
-func TestStatsTimeout(t *testing.T) {
-	tmpdir, err := ioutil.TempDir("", "socket")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.RemoveAll(tmpdir)
-	socketPath := filepath.Join(tmpdir, "docker_test.sock")
-	t.Logf("socketPath=%s", socketPath)
-	l, err := net.Listen("unix", socketPath)
-	if err != nil {
-		t.Fatal(err)
-	}
-	received := make(chan bool)
-	defer l.Close()
-	go func() {
-		conn, connErr := l.Accept()
-		if connErr != nil {
-			t.Logf("Failed to accept connection: %s", connErr)
-			return
-		}
-		breader := bufio.NewReader(conn)
-		req, connErr := http.ReadRequest(breader)
-		if connErr != nil {
-			t.Logf("Failed to read request: %s", connErr)
-			return
-		}
-		if req.URL.Path != "/containers/c/stats" {
-			t.Logf("Wrong URL path for stats: %q", req.URL.Path)
-			return
-		}
-		received <- true
-		time.Sleep(2 * time.Second)
-	}()
-	client, _ := NewClient("unix://" + socketPath)
-	client.SkipServerVersionCheck = true
-	errC := make(chan error, 1)
-	statsC := make(chan *Stats)
-	done := make(chan bool)
-	defer close(done)
-	go func() {
-		errC <- client.Stats(StatsOptions{ID: "c", Stats: statsC, Stream: true, Done: done, Timeout: time.Millisecond})
-		close(errC)
-	}()
-	err = <-errC
-	e, ok := err.(net.Error)
-	if !ok || !e.Timeout() {
-		t.Errorf("Failed to receive timeout error, got %#v", err)
-	}
-	recvTimeout := 2 * time.Second
-	select {
-	case <-received:
-		return
-	case <-time.After(recvTimeout):
-		t.Fatalf("Timeout waiting to receive message after %v", recvTimeout)
 	}
 }
